+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project **doesn't** adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) for now.
 
 ## [Unreleased]
+### Added
+- New "device" option for `FerpectionBadgeIcon` used in `ThematicSection`.
+
 ### Changed
 - We wrapp the `DatePickerField` with `UiKitInitializer` only on its storybook to display
   the documentation with translation support.